// src/app/models/year-end-inventory-models/unrestrictedZmmModel.ts
/**
 * year-end-inventory-backend
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UnrestrictedZmmModel { 
    material?: string;
    materialDescription?: string;
    plant?: string;
    storageLocation?: number;
    unrestricted?: number;
    baseUnitOfMeasure?: string;
    supplyingPlant?: string;
    specialStock?: string;
    salesDocument?: string;
    itemSD?: number;
    orderQuantity?: number;
    inQualityInsp?: number;
    blocked?: number;
    returns?: number;
    quantityReceived?: number;
    issuedQuantity?: number;
    purchasingDocument?: string;
    item?: number;
    materialDocument?: string;
    materialDocItem?: number;
    deliveryDate?: string;
    loadScheduleNumber?: string;
    deliveryStatus?: string;
}